requirements.txt
# API FastAPI pour la prédiction de matchs de football
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3

# Visualisation
matplotlib==3.8.2
seaborn==0.13.2

# Base de données
supabase==2.3.0

# Utilitaires
python-dotenv==1.0.1

